rename Array.reversed to Array.reverse
