+++ src/init.lua
@@ -40,7 +40,8 @@ local Disk = {
         removeIndexes = require('./array/removeIndexes'),
         removeSortedIndexes = require('./array/removeSortedIndexes'),
         removeValues = require('./array/removeValues'),
-        reversed = require('./array/reversed'),
+        reverse = require('./array/reverse'),
+        reversed = require('./array/reverse'),
         sort = require('./array/sort'),
         sortByKey = require('./array/sortByKey'),
         stepBy = require('./array/stepBy'),
